requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
python-dotenv
